[W] Finish Field Editor Changes
