+++ ClientService/ClientService.csproj
@@ -42,14 +42,19 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Brokers\EndListBroker.cs" />
+    <Compile Include="Brokers\BrokerSelector.cs" />
     <Compile Include="Brokers\SizeListBroker.cs" />
+    <Compile Include="Brokers\BrandListBroker.cs" />
     <Compile Include="CacheIO.cs" />
     <Compile Include="Brokers\SizeGroupBroker.cs" />
     <Compile Include="ContextProvider.cs" />
     <Compile Include="Contracts\IProvider.cs" />
+    <Compile Include="Data\EndProvider.cs" />
     <Compile Include="Data\CustomSizeProvider.cs" />
     <Compile Include="Data\SizeGroupProvider.cs" />
     <Compile Include="Data\SizeProvider.cs" />
+    <Compile Include="Data\BrandProvider.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SpecsManiuplator.cs" />
     <Compile Include="Data\SpecsProvider.cs" />
