+++ UserInterface/Forms/Main.cs
@@ -4,10 +4,11 @@
 //using CoreLibrary;
 //using CoreLibrary.Models;
 //using UserService;
-using CoreLibrary.Enums;
 using System;
 using System.Drawing;
 using System.Windows.Forms;
+
+using Modeling.DataModels;
 
 
 namespace UserInterface.Forms
@@ -36,17 +37,17 @@ namespace UserInterface.Forms
 
         private void SizeEditor_Click(object sender, EventArgs e)
         {
-            LauchEditor(new FieldEditor_<Modeling.DataModels.SizeList>());
+            LauchEditor(new FieldEditor<SizeList>());
         }
 
         private void BrandEditor_Click(object sender, EventArgs e)
         {
-            LauchEditor(new FieldEditor(FieldType.BRAND));
+            LauchEditor(new FieldEditor<BrandList>());
         }
 
         private void EndsEditor_Click(object sender, EventArgs e)
         {
-            LauchEditor(new FieldEditor(FieldType.ENDS));
+            LauchEditor(new FieldEditor<EndList>());
         }
 
         private void Exit_Click(object sender, EventArgs e)
@@ -96,8 +97,6 @@ namespace UserInterface.Forms
             Runtime.Test.AutoLoad(testLoadXmlContext);
             Runtime.Test.DoSomething(PostLoading);
             //Runtime.Test.DoSomething(delegate { LauchEditor(new SpecsEditor()); });
-            //Runtime.Test.DoSomething(delegate { LauchEditor(new SizeGroupEditor()); });
-            Runtime.Test.DoSomething(delegate { LauchEditor(new FieldEditor_<Modeling.DataModels.SizeList>()); });
         }
 
         private void tsmiLoadAll_Click(object sender, EventArgs e)
