--- UserInterface/Forms/XML Viewer.cs
@@ -1,14 +1,0 @@
-﻿
-using System.Windows.Forms;
-
-
-namespace UserInterface.Forms
-{
-    public partial class XML_Viewer : Form
-    {
-        public XML_Viewer()
-        {
-            InitializeComponent();
-        }
-    }
-}
